+++ test/IntegrationTests/WcfNetFrameworkTests.cs
@@ -19,7 +19,7 @@ public class WcfNetFrameworkTests : WcfTestsBase
     [Trait("Category", "EndToEnd")]
     public async Task SubmitsTraces()
     {
-        await SubmitsTracesInternal(string.Empty).ConfigureAwait(true);
+        await SubmitsTracesInternal(string.Empty, new WcfServerTestHelper(Output)).ConfigureAwait(true);
     }
 
     [Fact]