+++ test/IntegrationTests/WcfTestsBase.cs
@@ -46,6 +46,7 @@ public abstract class WcfTestsBase : TestHelper, IDisposable
                 else
                 {
                     _serverProcess.Process.Kill();
+                    _serverProcess.Process.WaitForExit();
                 }
 
                 Output.WriteLine("ProcessId: " + _serverProcess.Process.Id);
@@ -59,15 +60,22 @@ public abstract class WcfTestsBase : TestHelper, IDisposable
         _disposed = true;
     }
 
-    protected async Task SubmitsTracesInternal(string clientPackageVersion)
+    protected async Task SubmitsTracesInternal(string clientPackageVersion, WcfServerTestHelperBase wcfServerTestHelperBase)
     {
-        Assert.True(EnvironmentTools.IsWindowsAdministrator(), "This test requires Windows Administrator privileges.");
-
+        Assert.True(!EnvironmentTools.IsWindows() || EnvironmentTools.IsWindowsAdministrator(), "This test requires Windows Administrator privileges.");
+#if NET
+        Assert.NotNull(wcfServerTestHelperBase);
+#else
+        if (wcfServerTestHelperBase == null)
+        {
+            throw new ArgumentNullException(nameof(wcfServerTestHelperBase));
+        }
+#endif
         using var collector = new MockSpansCollector(Output);
         SetExporter(collector);
 
-        var serverHelper = new WcfServerTestHelper(Output);
-        _serverProcess = serverHelper.RunWcfServer(collector);
+        _serverProcess = wcfServerTestHelperBase.RunWcfServer(collector);
+
         await WaitForServer().ConfigureAwait(false);
 
         RunTestApplication(new TestSettings
@@ -75,9 +83,9 @@ public abstract class WcfTestsBase : TestHelper, IDisposable
             PackageVersion = clientPackageVersion
         });
 
-        collector.Expect("OpenTelemetry.Instrumentation.Wcf", span => span.Kind == SpanKind.Server, "Server 1");
+        collector.Expect(wcfServerTestHelperBase.ServerInstrumentationScopeName, span => span.Kind == SpanKind.Server, "Server 1");
         collector.Expect("OpenTelemetry.Instrumentation.Wcf", span => span.Kind == SpanKind.Client, "Client 1");
-        collector.Expect("OpenTelemetry.Instrumentation.Wcf", span => span.Kind == SpanKind.Server, "Server 2");
+        collector.Expect(wcfServerTestHelperBase.ServerInstrumentationScopeName, span => span.Kind == SpanKind.Server, "Server 2");
         collector.Expect("OpenTelemetry.Instrumentation.Wcf", span => span.Kind == SpanKind.Client, "Client 2");
 
         collector.Expect($"TestApplication.{_testAppName}", span => span.Kind == SpanKind.Internal, "Custom parent");