+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   SqlClient instrumentation setting
 - Add support for `StackExchange.Redis` for applications targeting .NET Framework.
 - Automatically set the `service.instance.id` resource attribute if not provided.
+- Support for [CoreWCF.Primitives](https://www.nuget.org/packages/CoreWCF.Primitives/)
+  traces instrumentation for versions `1.8.0`+.
 
 ### Changed